requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.2
django-simpleui==2025.5.262
django_celery_results==2.6.0
djangorestframework==3.16.0
flower==2.0.1
humanize==4.12.3
idna==3.10
kombu==5.5.4
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
